use mantine for app layout
