+++ components/header.tsx
@@ -1,12 +1,12 @@
 import Image from 'next/image';
-import { AppShell } from '@mantine/core';
+import { Group } from '@mantine/core';
 
 import logo from '../public/aclima-logo.svg';
 import classes from './header.module.css';
 
 export const Header: React.FC = () => {
   return (
-    <AppShell.Header className={classes.header}>
+    <Group>
       <Image src={logo} alt="Aclima logo" />
       <h1>Example Prototype</h1>
       <style jsx>{`
@@ -15,6 +15,6 @@ export const Header: React.FC = () => {
           margin: 0;
         }
       `}</style>
-    </AppShell.Header>
+    </Group>
   );
 };
