+++ components/prototype.module.css
@@ -1,0 +1,5 @@
+.map {
+  height: 100%;
+  width: 100%;
+  position: absolute;
+}
