+++ components/prototype.tsx
@@ -1,35 +1,23 @@
-import React from "react";
-import { Panel } from "./panel";
-import { Map } from "./map";
-import theme from "@/utils/theme";
-import { AmtrakResponse, BikeShareResponse } from "@/types";
+import React from 'react';
+import { AppShell } from '@mantine/core';
+import { PanelContents } from './panel';
+import { Map } from './map';
+import { AmtrakResponse, BikeShareResponse } from '@/types';
+import classes from './prototype.module.css';
 
 type MapDataResponses = {
   bikeShareData: BikeShareResponse;
   amtrakData: AmtrakResponse;
 };
-const MapContainer: React.FC<MapDataResponses> = ({bikeShareData, amtrakData}) => {
+export const MapContainer: React.FC<MapDataResponses> = ({
+  bikeShareData,
+  amtrakData,
+}) => {
   return (
-    <div>
-      <Map
-        bikeShareData={bikeShareData?.features ?? []}
-        amtrakData={amtrakData ?? { type: "FeatureCollection", features: [] }}
-      />
-      <style jsx>{`
-        div {
-          height: 500px;
-          width: 100%;
-          order: -1;
-          flex-shrink: 0;
-          @media (min-width: ${theme.breakpoints.mobile}) {
-            order: 0;
-            height: 100%;
-            flex-grow: 1;
-            width: auto;
-          }
-        }
-      `}</style>
-    </div>
+    <Map
+      bikeShareData={bikeShareData?.features ?? []}
+      amtrakData={amtrakData ?? { type: 'FeatureCollection', features: [] }}
+    />
   );
 };
 
@@ -39,19 +27,10 @@ export const Prototype: React.FC<MapDataResponses> = ({
 }) => {
   return (
     <div>
-      <Panel />
-      <MapContainer bikeShareData={bikeShareData} amtrakData={amtrakData}/>
-      <style jsx>{`
-        div {
-          height: 100%;
-          width: 100%;
-          display: flex;
-          flex-direction: column;
-          @media (min-width: ${theme.breakpoints.mobile}) {
-            flex-direction: row;
-          }
-        }
-      `}</style>
+      <PanelContents />
+      <AppShell.Main className={classes.main}>
+        <MapContainer bikeShareData={bikeShareData} amtrakData={amtrakData} />
+      </AppShell.Main>
     </div>
   );
 };
