+++ pages/index.module.css
@@ -1,3 +1,9 @@
+.main {
+  height: 100%;
+  width: 100%;
+  position: absolute;
+}
+
 .header {
   display: flex;
   align-items: center;
